requirements.txt
python-telegram-bot
fritzconnection
schedule
SQLAlchemy
pymysql